fix: allow fractional cost_in_credits in JobResponse schema
Decimal('907.5') from DB caused Pydantic validation error since
cost_in_credits was typed as int. Changed to float.

Made-with: Cursor

+++ backend/app/schemas/job.py
@@ -16,7 +16,7 @@ class JobResponse(BaseModel):
     processed_leads: int
     valid_emails_found: int
     catchall_emails_found: int
-    cost_in_credits: int
+    cost_in_credits: float
     input_file_path: Optional[str]
     output_file_path: Optional[str]
     created_at: datetime
